refactor(studio): move two composition pages out of features, 10 -> 6 [GH-000]
studio had three cross-feature imports and all three were in one file.
ProductListPage sat in features/products and imported the pending-order
count from features/orders and the delegate counts and a permission
constant from features/seller-admins.

The imports were not the problem. A page that needs three features is a
composition root, and a feature is the wrong place for one. Moved to
shared/presentation/pages, where composing several features is what the
layer is for; the route still renders it and nothing else, as the rule
asks.

ProductDelegatesPage is the same shape -- it manages delegates and needs
the product record to title the page -- and moved for the same reason.

studio now has no cross-feature imports at all.

Six remain, and they are two pairs that import each other:

  payments  assigned-orders and received-orders, 4. Assigned orders are
            received orders seen through a delegation lens: the same
            card, the same actions hook, the same SellerAction type, and
            assigned-orders' own query function even lives in
            received-orders' file. Resolving it means lifting the shared
            order-management domain into shared/ and leaving two thin
            features that differ only in which orders they list. That is
            a design decision, and moving fetchAssignedOrders on its own
            would only trade an import for a type import.

  store     cart and products, 2. A product card adds to the cart; the
            cart needs product records to render. Genuinely mutual.

Both are recorded at the top of the check script with the resolution I
would suggest. Neither is a lint fix, so neither was done by one.

Baseline 10 -> 6. lint 0, typecheck 12/12, studio 416 passing.
