+++ apps/studio/src/app/[locale]/products/[id]/delegates/page.tsx
@@ -1,6 +1,6 @@
 import { setRequestLocale } from "next-intl/server";
 
-import { ProductDelegatesPage } from "@/features/seller-admins";
+import { ProductDelegatesPage } from "@/shared/presentation/pages/ProductDelegatesPage";
 
 export default async function ProductDelegatesRoute({
   params,
