+++ apps/studio/src/features/products/index.ts
@@ -1,3 +1,2 @@
 export { ProductFormPage } from "./presentation/pages/ProductFormPage";
-export { ProductListPage } from "./presentation/pages/ProductListPage";
 export { useProductById } from "./application/hooks/useProductForm";
