+++ apps/studio/src/features/seller-admins/index.ts
@@ -1,6 +1,5 @@
 // Pages
 export { DelegateManagementPage } from "./presentation/pages/DelegateManagementPage";
-export { ProductDelegatesPage } from "./presentation/pages/ProductDelegatesPage";
 
 // Hooks
 export { useDelegates } from "./application/hooks/useDelegates";
