+++ scripts/check-feature-boundaries.mjs
@@ -6,11 +6,28 @@
  * import directly from other features", with cross-feature communication going
  * through shared interfaces, stores, or events. There are 13 that do.
  *
- * This is a ratchet, not a clean-up. Each of those 13 needs a decision that
- * is bigger than an import statement -- three of the pairs import each other,
- * which usually means they are one feature wearing two names -- and making
- * that call is not something a lint sweep should do. What this prevents is a
- * fourteenth arriving while nobody is looking.
+ * This is a ratchet. It started at 13; seven had a remedy the rule already
+ * prescribes and are gone -- query keys and audit writing moved to shared, and
+ * two pages that composed three features each moved out of the feature they
+ * happened to live in.
+ *
+ * The remaining six are two pairs that import each other, and they need a
+ * decision bigger than an import statement:
+ *
+ *   payments  assigned-orders and received-orders, 4 imports. Assigned orders
+ *             are received orders seen through a delegation lens: the same
+ *             card, the same actions hook, the same SellerAction type, and
+ *             assigned-orders' own query function even lives in
+ *             received-orders' file. Resolving it means lifting the shared
+ *             order-management domain into shared/ and leaving two thin
+ *             features that differ only in which orders they list.
+ *
+ *   store     cart and products, 2 imports. A product card needs to add to the
+ *             cart; the cart needs product records to render. Genuinely
+ *             mutual, and the cheapest honest fix is probably to invert one
+ *             direction with a callback prop.
+ *
+ * Neither is a lint fix, so neither was done by one.
  *
  * It also replaced a claim in docs/standards/quality-gates.md that the feature
  * barrel rule "is inconsistent with itself", evidenced by 126 deep imports
@@ -30,7 +47,7 @@ import { readFileSync } from "node:fs";
  * What the codebase had when this check was written. It may fall; it may not
  * rise. Lower it when a pair is resolved -- that is the ratchet.
  */
-const BASELINE = 10;
+const BASELINE = 6;
 
 const FEATURE_IMPORT = /from\s+"(@\/features\/([^/"]+)([^"]*))"/g;
